+ Improved auth logic (look for Breaking Changes below) + Improved logging + Improved app data handling to support unified Int64 value format for identifiers

+++ ThunderED/Classes/DiscordCommands.cs
@@ -228,7 +228,7 @@ namespace ThunderED.Classes
                 try
                 {
                     string code;
-                    int characterId;
+                    long characterId;
                     if (!data.All(char.IsDigit))
                     {
                         code = data;
@@ -237,7 +237,7 @@ namespace ThunderED.Classes
                     else
                     {
                         code = null;
-                        characterId = Convert.ToInt32(data);
+                        characterId = Convert.ToInt64(data);
                     }
                     code = code ?? await SQLHelper.PendingUsersGetCode(characterId);
