+++ ThunderED/Json/FleetUp/JsonFleetup.cs
@@ -17,25 +17,25 @@ namespace ThunderED.Json.FleetUp
 
         public class Datum
         {
-            public int Id { get; set; }
-            public int OperationId { get; set; }
+            public long Id { get; set; }
+            public long OperationId { get; set; }
             public string Subject { get; set; }
             public DateTime Start { get; set; }
             public DateTime End { get; set; }
             public string Location { get; set; }
-            public int LocationId { get; set; }
+            public long LocationId { get; set; }
             public string Details { get; set; }
             public string Url { get; set; }
             public string Organizer { get; set; }
             public string Category { get; set; }
             public string Group { get; set; }
-            public int GroupId { get; set; }
+            public long GroupId { get; set; }
             public Doctrine[] Doctrines { get; set; }
         }
 
         public class Doctrine
         {
-            public int Id { get; set; }
+            public long Id { get; set; }
             public string Name { get; set; }
         }
     }
